# backend/internal/adapters/http/dto/dto.go
// Package dto contains Data Transfer Objects for HTTP request/response in Kodia Framework.
// DTOs validate and sanitize input before passing it to the service layer.
package dto

// --- Auth DTOs ---

// RegisterRequest is the request body for POST /api/auth/register.
// @swagger:model
type RegisterRequest struct {
	// User display name
	// @required
	// @example "John Doe"
	Name string `json:"name" validate:"required,min=2,max=100" example:"John Doe"`

	// User email address
	// @required
	// @example "user@example.com"
	Email string `json:"email" validate:"required,email" example:"user@example.com"`

	// User password
	// @required
	// @example "SecurePassword123!"
	// @minLength 8
	Password string `json:"password" validate:"required,min=8,max=72" example:"SecurePassword1123!"`
}

// LoginRequest is the request body for POST /api/auth/login.
// @swagger:model
type LoginRequest struct {
	// User email address
	// @required
	// @example "user@example.com"
	Email string `json:"email" validate:"required,email" example:"user@example.com"`

	// User password
	// @required
	Password string `json:"password" validate:"required" example:"SecurePassword123!"`
}

// RefreshTokenRequest is the request body for POST /api/auth/refresh.
// @swagger:model
type RefreshTokenRequest struct {
	// Valid refresh token
	// @required
	RefreshToken string `json:"refresh_token" validate:"required" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."`
}

// LogoutRequest is the request body for POST /api/auth/logout.
// @swagger:model
type LogoutRequest struct {
	// Refresh token to revoke
	// @required
	RefreshToken string `json:"refresh_token" validate:"required" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."`
}

// AuthResponse is the response for successful authentication.
// @swagger:model
type AuthResponse struct {
	AccessToken  string       `json:"access_token"  example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."`
	RefreshToken string       `json:"refresh_token" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."`
	TokenType    string       `json:"token_type"    example:"Bearer"`
	User         UserResponse `json:"user"`
	MFARequired  bool         `json:"mfa_required,omitempty"`
	MFAToken     string       `json:"mfa_token,omitempty"`
}

// ForgotPasswordRequest is the request body for POST /api/auth/forgot-password.
// @swagger:model
type ForgotPasswordRequest struct {
	// User email address
	// @required
	Email string `json:"email" validate:"required,email" example:"user@example.com"`
}

// ResetPasswordRequest is the request body for POST /api/auth/reset-password.
// @swagger:model
type ResetPasswordRequest struct {
	// Reset token from email
	// @required
	Token string `json:"token" validate:"required" example:"550e8400-e29b-41d4-a716-446655440000"`

	// New password
	// @required
	NewPassword string `json:"new_password" validate:"required,min=8,max=72" example:"NewSecurePassword456!"`
}

// Verify2FARequest is the request body for POST /api/auth/2fa/verify.
// @swagger:model
type Verify2FARequest struct {
	// 6-digit TOTP code
	// @required
	Code string `json:"code" validate:"required,len=6" example:"123456"`
}

// TwoFactorSetupResponse is the response for enabling 2FA.
// @swagger:model
type TwoFactorSetupResponse struct {
	Secret string `json:"secret" example:"JBSWY3DPEHPK3PXP"`
	QRCode string `json:"qr_code" example:"otpauth://totp/Kodia%20Framework:user@example.com?secret=JBSWY3DPEHPK3PXP&issuer=Kodia%20Framework"`
}

// LoginVerify2FARequest is the request body for POST /api/auth/2fa/login-verify.
// @swagger:model
type LoginVerify2FARequest struct {
	// Temporary MFA token from login response
	// @required
	MFAToken string `json:"mfa_token" validate:"required" example:"550e8400-e29b-41d4-a716-446655440000"`

	// 6-digit TOTP code
	// @required
	Code string `json:"code" validate:"required,len=6" example:"123456"`
}

// --- User DTOs ---

// UserResponse is the public-safe user representation.
// Never expose the password field.
// @swagger:model
type UserResponse struct {
	// User unique identifier
	ID string `json:"id" example:"550e8400-e29b-41d4-a716-446655440000"`

	// User display name
	Name string `json:"name" example:"John Doe"`

	// User email address
	Email string `json:"email" example:"user@example.com"`

	// User role
	Role string `json:"role" example:"user"`

	// User account status
	IsActive bool `json:"is_active" example:"true"`

	// Email verification status
	IsVerified bool `json:"is_verified" example:"false"`

	// 2FA status
	TwoFactorEnabled bool `json:"two_factor_enabled" example:"false"`

	// User avatar image URL
	AvatarURL *string `json:"avatar_url" example:"https://example.com/avatar.jpg"`

	// ISO 8601 creation timestamp
	CreatedAt string `json:"created_at" example:"2024-04-19T10:00:00Z"`

	// ISO 8601 update timestamp
	UpdatedAt string `json:"updated_at" example:"2024-04-19T10:00:00Z"`
}

// UpdateUserRequest is the request body for PATCH /api/users/:id.
// @swagger:model
type UpdateUserRequest struct {
	Name      *string `json:"name"       validate:"omitempty,min=2,max=100" example:"Jane Doe"`
	AvatarURL *string `json:"avatar_url" validate:"omitempty,url"             example:"https://example.com/avatar2.jpg"`
}

// ChangePasswordRequest is the request body for POST /api/users/me/change-password.
// @swagger:model
type ChangePasswordRequest struct {
	CurrentPassword string `json:"current_password" validate:"required" example:"OldSecurePassword123!"`
	NewPassword     string `json:"new_password"     validate:"required,min=8,max=72" example:"NewSecurePassword456!"`
}

// PaginatedUsersResponse wraps a list of users with pagination metadata.
// @swagger:model
type PaginatedUsersResponse struct {
	Users []UserResponse `json:"users"`
}

// --- Notification DTOs ---

// NotificationResponse is the response for a notification.
// @swagger:model
type NotificationResponse struct {
	ID        string                 `json:"id" example:"550e8400-e29b-41d4-a716-446655440000"`
	Type      string                 `json:"type" example:"info"`
	Title     string                 `json:"title" example:"Welcome!"`
	Message   string                 `json:"message" example:"Welcome to Kodia Framework"`
	Data      map[string]interface{} `json:"data,omitempty"`
	IsRead    bool                   `json:"is_read" example:"false"`
	CreatedAt string                 `json:"created_at" example:"2024-04-19T10:00:00Z"`
}

// UnreadCountResponse is the response for unread notification count.
// @swagger:model
type UnreadCountResponse struct {
	Count int64 `json:"count" example:"5"`
}

// --- Role DTOs ---

// CreateRoleRequest is the request body for POST /api/admin/roles.
// @swagger:model
type CreateRoleRequest struct {
	// Role name
	// @required
	Name string `json:"name" validate:"required,min=1,max=255" example:"editor"`

	// Role description
	Description string `json:"description" example:"Can edit content"`

	// List of permission names
	Permissions []string `json:"permissions" example:"content:read,content:write"`
}

// RoleResponse is the response for a role.
// @swagger:model
type RoleResponse struct {
	// Role unique identifier
	ID string `json:"id" example:"550e8400-e29b-41d4-a716-446655440000"`

	// Role name
	Name string `json:"name" example:"editor"`

	// Role description
	Description string `json:"description" example:"Can edit content"`

	// List of permission names
	Permissions []string `json:"permissions" example:"content:read,content:write"`

	// ISO 8601 creation timestamp
	CreatedAt string `json:"created_at" example:"2024-04-19T10:00:00Z"`

	// ISO 8601 update timestamp
	UpdatedAt string `json:"updated_at" example:"2024-04-19T10:00:00Z"`
}

// AssignRoleRequest is the request body for POST /api/admin/users/:user_id/roles.
// @swagger:model
type AssignRoleRequest struct {
	// Role name to assign
	// @required
	RoleName string `json:"role_name" validate:"required,min=1" example:"editor"`
}